+++ Backend/BirdAPI.Domain/AggregatesModel/OwnerAggregate/OwnerBird.cs
@@ -5,9 +5,9 @@ namespace BirdAPI.Domain.AggregatesModel.OwnerAggregate
 {
     public class OwnerBird : EntityLong
     {
-        public long OwnerId { get; set; }
+        public int OwnerId { get; set; }
         public virtual Owner Owner { get; set; }
-        public long BirdId { get; set; }
+        public int BirdId { get; set; }
         public virtual Bird Bird { get; set; }
 
         public OwnerBird()
